requirements.txt
whisper
ffmpeg-python
moviepy
flask